Fix: AI와의 API 연동 이슈 수정

See merge request s11-s-project/S11P21S002!195

+++ backend/src/main/java/com/worlabel/domain/project/dto/AiDto.java
@@ -54,7 +54,7 @@ public class AiDto {
 
         private int epochs; // Default = 50
 
-        private double batch; // Default = -1
+        private int batch; // Default = -1
 
         private double lr0;
